+++ hls-polybench/seidel-2d/seidel-2d_tb.cpp
@@ -31,7 +31,7 @@ void print_array(int n,
   for (i = 0; i < n; i++)
     for (j = 0; j < n; j++) {
       if ((i * n + j) % 20 == 0) fprintf(stderr, "\n");
-      fprintf(stderr, "%0.2lf ", A[i][j]);
+      fprintf(stderr, "%0.6lf ", (float)A[i][j]);
     }
   fprintf(stderr, "\nend   dump: %s\n", "A");
   fprintf(stderr, "==END   DUMP_ARRAYS==\n");
